+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.0 - Nov 10, 2015
 * Added `seqa` generator.
-** Same as `seq` but takes an array of generators.
+** Same as `seq` but takes an array-ish of generators.
 * Use array literals to express sequences in grammars:
 ** `pep.choice(['a', 3], ['b', 4])` Yields either 'a' then '3', or 'b' then '4'.